refactor(game-card): extract game board logic to custom hook
Move card styling and interaction logic from GameCard component to useGameBoard hook
Add utility functions for gradients and shadows in game-board utils
